version 2.048 - June 2011
1. When searching for unique merged bam files, the sample name and
technology appearing in multiple files causes a problems, as the
pipeline is unclear as to which existing bam file is the correct
file to carry forward.  In checking the names, the full path was
included, so if the technology/sample name appeared in a directory
name, an error was thrown and the pipeline terminated.  This has
been modified to only include the file name itself.