version 2.057 - August 2011
1. Added the command -snp-delimiter.  When SNP calling, target regions
of the form <chr>:<start>-<end> are created and are used in the command
line for some of the software tools.  The region also appears in the
filename.  Some systems do not like having ":" in the filename, so
using this command allows the user to specify a different character.
The software tools will still be given the expected region format.