version 2.058 - September 2011
1. Added the -scratch <dir> option to the command line.  The default
location on the nodes is /scratch/<user>, but specifying this
command will change the directory to /<dir>/<user>.

2. Added the -fastq-list <list> option to the command line.  Instead
of searching for and checking that fastq files exist, the list of
fastq files in the specified list will be used.  The scripts will use
the path specified, so it is the users responsibility to ensure that
these files exist.